Geral  -Alteração em varios daos trocando o datatype de int pra float nos itens que tratam de medição. Esse datatype foi definido errado na documentação e quem dançou fui que tive que refazer. @#$ GE_Custom_Form_DataDao, GE_Custom_Form_Data , GeOsDao , GeOs , DatabaseHelperMulti  - Modificado Measure_cycle_value de int para float
@#$
DatabaseHelperMulti , MeMeasureTpDao e MeMeasureTp
 - Modificado value_cycle_size de int para float
@#$
MD_Product_Serial, MD_Product_SerialDao e DatabaseHelperMulti, MD_Product_Serial_Structure
 - Alterado datatype de Last_cycle_value de int para float
@#$
GeOs, GeOsDao e DatabaseHelperMulti
 - Alterado datatype de Last_cycle_value de int para float
 - Add os campos VALUE_CYCLE_SIZE e CYCLE_TOLERANCE de MeMeasure para identificar se a medição é ciclica ou não
 - Em GeOsDao, modificado fun checkScan para que sempre seja usado o valor de measure_cycle_value, pois agora como o datattype é o mesmo measure_value, quando a medição não for preventiva ciclica o valor de  measure_value será copiado para measure_cycle_value. Quando for uma preventiva ciclica, o valor será diferente.
@#$
Act011_Main_Presenter_Impl
 - Alterado metodo loadAnswer, atualizando logica que identifica se é uma preventiva ciclica para setar ou não o valor de setMeasure_cycle_value no obj form data
@#$
Act087MainPresenter.kt
 - Atualizado criação do obj GeOs com os novos campos
@#$
FormOsHeaderFrg.kt
 - Alterado datatype da proprieade calculatedExecCycle de int para float
 - Modificado inicialização da proprieade
 - Criado fun isPreventiveCycledOs para identificar se se trata de uma preventiva ciclica.
 - Alterado fun getMeasureCycleValue, para caso seja um preventiva ciclica, devolver o valor de calculatedExecCycle senão, devolve o mesmo valor de measure_value. Alteração feita para facilitar as trataivas nas varredura pois analisa sempre o mesmo campo. Quando não e um preventiva ciclica o valor de measure_cycle_value não serve pra nada.
 - Alterado fun isPreventiveCycleValid, extraindo a logica que identicia se é preventiva ciclicada para a fun isPreventiveCycledOs. Alterado inicialização de lastCycle, que teve seu datatype alterado para float. Removido toInt dao calculo de fatorNextCycle, pois ele pode ser float.
 - Alterado inicialização dos param s  calculatedCycle e lastCycleVal na assinatra da fun showSaveErroDialog
 -

+++ app/src/main/java/com/namoadigital/prj001/dao/MD_Product_SerialDao.java
@@ -1115,7 +1115,7 @@ public class MD_Product_SerialDao extends BaseDao implements Dao<MD_Product_Seri
             if(cursor.isNull(cursor.getColumnIndex(LAST_CYCLE_VALUE))){
                 md_product_serial.setLast_cycle_value(null);
             }else{
-                md_product_serial.setLast_cycle_value(cursor.getInt(cursor.getColumnIndex(LAST_CYCLE_VALUE)));
+                md_product_serial.setLast_cycle_value(cursor.getFloat(cursor.getColumnIndex(LAST_CYCLE_VALUE)));
             }
 
             //
